requirements.txt
flask-SQLAlchemy
flask
flask-cors
cheroot